Add in crossword word selection code
This is taken from the experimental project:
https://github.com/animatinator/experimental/tree/master/crossword-dictionary

Just drop the code in for now. In future, this will be combined with
the surrounding crossword generation code to generate puzzles for
the game.

An omission here is how dictionaries will be stored and loaded now
that we're on Android. The basic ones can be assets/raw resources,
but it might be good to be able to download dictionaries
dynamically. This will be addressed later.