Start adding a PuzzleGenerator
This will generate a PuzzleConfiguration, which contains both the word
config and the board layout.

PuzzleGenerator is only one-third-done at the moment as it just sets up
the word configuration; subsequent changes will flesh it out.

This also adds a PuzzleGenerationSettings class which contains the
preferences for generating the puzzle.