Use different hit radii for initial keyboard clicks versus drags
The initial click should use a more permissive radius, but subsequent
ones should be tight to avoid accidental hits whilst dragging. This
seems to be the right balance to make the keyboard usable - the last
change made dragging mistakes less likely, but meant it was occasionally
hard to tap the letter in the first place.