Add a Sodding Gargantuan English dictionary
This is probably too much - the puzzles are now quite hard and use some
very obscure words, with vast numbers of bonus words.

A good future direction would be a medium-sized dictionary with a giant
dictionary for bonus word lookup - that way the generation only uses
simpler words but all words will be recognised as bonus words.