Extract evaluation of word configs into WordConfigurationEvaluator
This will make it easier to swap it out for other means of evaluation -
now that we have a bigger generation dictionary, just picking the base
word which can generate the most words means a massive number of bonus
words and a frustrating puzzle. We might want to switch to more nuanced
ways of comparing configs.