Add more nuanced evaluation of word configurations
Instead of maximising possible words, we should try to pick a base word
which generates as close as possible to the right number of other words.

We intentionally punish too few words far more harshly than too many, so
we don't end up generating tiny puzzles over normal-sized ones with lots
of bonus words. This might not be the best plan: maybe instead we should
try to minimise bonus words as much as possible? That's TBC though.