Fix padding for EnteredTextDisplay and fix text centring in general
Turns out Paint.getTextBounds is a pretty solid approach to getting the
painted size of a string, so use that and remove the hack in
TextDrawingUtils (which turned out to be returning negative numbers!
How did it even work?!).

Given that more reliable measure, make the horizontal padding for the
text entry display the same as the vertical padding (don't just make it
a fraction of the horizontal text size because that makes it grow as the
string grows, which is strange).