Add infrastructure for monitoring the progress of board generation
We'll use this in switching generation to an interstitial with a
progress bar.

Add ProgressListener, an interface allowing reporting progress from 0.0d
to 1.0d. Then take it as an argument in word config and board layout
generation and implement appropriately.

Also add PuzzleGenerationProgressCallback, an extension interface which
adds the ability to describe the current work being done, and implement
appropriately for PuzzleGenerator (combining the progress of the two
sub-stages with relative fractions of the total progress defined in
constants).