Move puzzle generation to a background thread
In the same activity for now. Instead of generating the puzzle on the
main thread, we use an ExecutorService to generate it in the background
and then set it on the game view.

Currently this means we jump straight into the game activity with a
dummy puzzle layout (which will crash on contact!), only to replace it
after a couple of seconds with the real generated layout.

A follow-up will add an interstitial activity which generates the
puzzle, shows a progress bar etc.