Add a means of identifying when a puzzle is complete
This just naively scans through all the present board positions to see
if they're all complete. The boards are all pretty small so we don't
seem to need anything more efficient.

Also fix and add a load of tests of calculateSize to CrosswordLayoutTest
since it was causing some problems in testing the new isFinished method.